corrijo errores, sigue colgandose en el join y al meter a un segundo cliente me diceThe queue is already closed.

+++ server_sender_thread.cpp
@@ -8,11 +8,17 @@ void SenderThread::run() {
     ServerProtocol protocol(client_skt);
     while (connection_alive && still_alive) {
         ServerMessage message = server_messages.pop();
-        protocol.send_status(message);
         connection_alive = not protocol.get_was_closed();
+        if (connection_alive) {
+            //quizas chequear que se envíe solamente cuando hay un mensaje efectivamente
+            protocol.send_status(message);
+        } else {
+            std::cout << "connection alive del sender es false" << std::endl;
+            queues.delete_queue(id);
+            server_messages.close();
+        }
     }
-    queues.delete_queue(id);
-    server_messages.close();
+    std::cout << "Sale del while del sender" << std::endl;
     /*
     if (!connection_alive) {
         queues.delete_queue(id);
